g4-common: change /misc to grow (READ!)
long story short: this fixes OTA on UsU'd devices
(there's also a "TL;DR" at the bottom)

*WARNING WARNING WARNING WARNING*

!! This change requires a compatible TWRP !!

details about the current (i.e. previous) implementation:
https://github.com/steadfasterX/android_device_lge_g4/commit/d8053c20a97607d

this all works perfectly fine until.. an UsU (!) device is trying to do an OTA
upgrade. Something within aboot removes the information from within
/misc !! weird? yes, weird!!
I did a LOT of tests to find out where and how that happens but the only
thing left is aboot - and it WORKS when doing the following:

*WORKAROUND*
--------------------
1. apply OTA and watch briefly after pressing INSTALL
2. as soon as the device shuts down: take out the battery
3. power on and let it boot -> TWRP will pop-up and the OTA will be
applied.

*REASON*
--------------------
So why? the reboot -> recovery command does a warm reset of the
device and during that warm reset the removal of the commands in /misc
happen. On the other hand a power off + power on (cold restart) will
NOT remove these and so works as expected. Which means UsU aboot is
doing nasty things (hey its a debug aboot so we shouldn't be that rude).

We can have only one /misc and non UsU'd devices are not affected but
have another thing to keep in mind: the LG factory screen.
The hw key combo to factory reset is not available on UsU'd devices.
Non-UsU'd devices needs to walk through this menu in order to boot
recovery though.

As stated at the top:
!! This change requires a compatible TWRP !!

*COMPATIBILITY*
--------------------
for custom ROMs including this commit:
- TWRP >= build 231
- SHRP >= build 54

for STOCK ROMs or ROM's which do not have this commit:
- TWRP =< build 230
- SHRP =< build 53

*SCENARIOS*
--------------------
What happens when you do NOT have a compatible combo between
recovery and ROM and you do an OTA or factory reset from within
Android?

Here are the possible scenarios:

OLD recovery version, ROM does NOT have this  commit:
- UsU device -> will just boot recovery
- non-UsU device -> will work as expected (OTA upgrade, fac reset)

NEW recovery version but ROM does NOT have this  commit:
- UsU device -> recovery boot loop
- non-UsU device -> recovery boot loop

OLD recovery version, ROM does have this commit:
- UsU device -> recovery boot loop
- non-UsU device -> recovery boot loop

NEW recovery version, ROM does have this  commit:
- UsU device -> will work as expected (OTA upgrade, fac reset)
- non-UsU device -> will work as expected (OTA upgrade, fac reset)

NEW recovery version, STOCK ROM:
- UsU device -> recovery boot loop
- non UsU device -> recovery boot loop

OLD recovery version, STOCK ROM:
- UsU device -> will just boot recovery
- non UsU device -> will just boot recovery (fac reset, OTA? lol!)

*UPGRADE PATH*
--------------------
with OTA:
1. wait until an OTA comes out which includes this commit
2. follow the steps from above: *WORKAROUND*
3. upgrade your recovery (see above: *COMPATIBILITY*)

without OTA:
1. wait until your fav ROM provides a version which includes this commit
2. upgrade to that new ROM release
3. upgrade your recovery (see above: *COMPATIBILITY*)

when on STOCK:
- stay on a compatible recovery version (see above: *COMPATIBILITY*)!

*TL;DR*
--------------------
1) this introduces the previous "grow" as /misc again.This way we should
be safe for both types (UsU or not, factory reset screen or direct).

2) the already known thing stays:
a factory reset from Android will just boot to recovery.
Not a big deal -> just do a factory reset from within the recovery then.

3) ensure you have a compatible recovery installed otherwise = bootloop!
 (see above: *COMPATIBILITY*)

Change-Id: Ic5142357b98c1b1fdc03fbbd56a60cc5d234c17f