g4-common: disable recovery flash option
when a ROM does not contain a recovery and the user had the option
enabled to update the recovery during a ROM flash the recovery will be
ZEROED out! IMHO a serious bug by LOS(?).

to prevent that we:

1. reset the corresponding property on every reboot
2. remove the option to update recovery from Updater

the user can still override this by setting manually:

`setprop persist.vendor.recovery_update true`

and then directly flashing.

Change-Id: Ic1dde80c9cb36265adc5e20878a2487373ee7a15