g4-common: respect gps.conf for SUPL_HOST and PORT
not sure where exactly the issue comes from but when you set
SUPL_HOST or SUPL_PORT within the gps.conf it "should" overwrite
any defaults coming from android.
this is hard to track and .... simply NOT the case (for the lg g4 at least).

this commit ensures that the gps.conf will win always - not the android default.

Note: which SUPL_HOST / _PORT is used can be checked by the boot
logs ONLY. Any attempt to see that later on when Android is running already
will not show the corresponding log entries as this setup happens on boot only.

Signed-off-by: steadfasterX <steadfasterX@gmail.com>